Remove unnecessary static_cast

+++ libethcore/TransactionBase.cpp
@@ -76,7 +76,7 @@ TransactionBase::TransactionBase(bytesConstRef _rlpData, CheckTransaction _check
 		else
 		{
 			if (v > 36)
-				m_chainId = static_cast<int>((v - 35) / 2); 
+				m_chainId = (v - 35) / 2; 
 			else if (v == 27 || v == 28)
 				m_chainId = -4;
 			else
